refactor: Extract taxonomy methods to view model
